feat: 实现完整的个人主页与安全登录系统
添加静态资源文件(图片、字体)和.gitignore配置
重构app.py实现密码哈希存储和首次登录检测
创建响应式HTML模板和CSS样式
添加JavaScript实现泡泡动画和响应式布局
