+++ requirements.txt
@@ -1,3 +1,4 @@
 Flask==2.0.1
 Werkzeug==2.0.1
 Jinja2==3.0.1
+bcrypt==3.2.0
